Ticket 1422: EndianProtocol needed to changed to size_t to compile in 64 bit GDC

+++ tango/io/protocol/EndianProtocol.d
@@ -84,7 +84,7 @@ class EndianProtocol : NativeProtocol
                 
                 void write (int mask, Swapper mutate)
                 {
-                        uint writer (void[] dst)
+                        size_t writer (void[] dst)
                         {
                                 // cap bytes written
                                 uint len = dst.length & mask;
